update cutom model
