+++ qllm/__init__.py
@@ -1,2 +1,5 @@
 from . import run
+
+__version__ = '0.1.2'
+
 __all__ = ['run']
