--- versions.txt
@@ -1,2 +1,0 @@
-__version__ = "0.1.2"
-
